fixed bugs in user controller and cloudinary

+++ src/utils/cloudinary.js
@@ -1,21 +1,36 @@
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
-const uploadOnCloudinary = async (localFilePath)=>{
+
+const configureCloudinary = () => {
+  cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
+  });
+//   console.log("🌩️ Cloudinary configured:", cloudinary.config());
+};
+
+const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath) return null;
-
-        const response = await cloudinary.uploader.upload(localFilePath,{
-            resource_type:"auto"
+        if (!localFilePath) return null
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto"
         })
+        // console.log("file is uploaded on cloudinary",response.url);
 
-        console.log("File uploaded successfully",response.url);
-        return response
-        
+        fs.unlinkSync(localFilePath)
+        return response;
+
     } catch (error) {
-        fs.unlinkSync(localFilePath);
-        return null;
+        console.log("cloudinary upload error",error);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath)
+        }
+        return null
     }
 }
 
-export {uploadOnCloudinary}
+
+
+export {configureCloudinary,uploadOnCloudinary}
